requirements.txt
# openai==1.1.1
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
python-dotenv==1.0.1
pinecone-client==2.2.4
tiktoken==0.5.2
streamlit==1.30.0